We have worked on the autonomous.
We are doing a step  by step autonomous that can be altered quickly to
fit one of the three different positions at a given time. We also have
corrected the plunger movements per sec. The shooter motor itself and
tested the navx to be able to use its functions while the axis is
altered.